Make tests easier to debug
