+++ pkg/server/last_operation_test.go
@@ -85,7 +85,7 @@ func TestLastOperation(t *testing.T) {
 			fs := httptest.NewServer(s.Router)
 			defer fs.Close()
 
-			config := osb.DefaultClientConfiguration()
+			config := defaultClientConfiguration()
 			config.URL = fs.URL
 
 			client, err := osb.NewClient(config)
