+++ pkg/server/provision_test.go
@@ -98,7 +98,7 @@ func TestProvision(t *testing.T) {
 			fs := httptest.NewServer(s.Router)
 			defer fs.Close()
 
-			config := osb.DefaultClientConfiguration()
+			config := defaultClientConfiguration()
 			config.URL = fs.URL
 
 			client, err := osb.NewClient(config)
