+++ pkg/server/server_test.go
@@ -1,14 +1,9 @@
 package server
 
 import (
-	"errors"
 	"net/http"
-	"net/http/httptest"
-	"reflect"
-	"testing"
 
 	"github.com/pmorie/osb-starter-pack/pkg/broker"
-	"github.com/pmorie/osb-starter-pack/pkg/rest"
 
 	osb "github.com/pmorie/go-open-service-broker-client/v2"
 )
@@ -70,82 +65,9 @@ func strPtr(s string) *string {
 	return &s
 }
 
-func TestGetCatalog(t *testing.T) {
-	okResponse := &osb.CatalogResponse{Services: []osb.Service{
-		{
-			Name: "foo",
-		},
-	}}
+func defaultClientConfiguration() *osb.ClientConfiguration {
+	conf := osb.DefaultClientConfiguration()
+	conf.Verbose = true
 
-	cases := []struct {
-		name         string
-		validateFunc func(string) error
-		catalogFunc  func(w http.ResponseWriter, r *http.Request) (*osb.CatalogResponse, error)
-		response     *osb.CatalogResponse
-		err          error
-	}{
-		{
-			name: "OK",
-			catalogFunc: func(w http.ResponseWriter, r *http.Request) (*osb.CatalogResponse, error) {
-				return okResponse, nil
-			},
-			response: okResponse,
-		},
-		{
-			name: "version validation error",
-			validateFunc: func(string) error {
-				return errors.New("oops")
-			},
-			err: osb.HTTPStatusCodeError{
-				StatusCode:  http.StatusPreconditionFailed,
-				Description: strPtr("oops"),
-			},
-		},
-	}
-
-	for i := range cases {
-		tc := cases[i]
-		t.Run(tc.name, func(t *testing.T) {
-			validateFunc := defaultValidateFunc
-			if tc.validateFunc != nil {
-				validateFunc = tc.validateFunc
-			}
-
-			api := &rest.APISurface{
-				BusinessLogic: &FakeBusinessLogic{
-					validateAPIVersion: validateFunc,
-					getCatalog:         tc.catalogFunc,
-				},
-			}
-
-			s := New(api)
-			fs := httptest.NewServer(s.Router)
-			defer fs.Close()
-
-			config := osb.DefaultClientConfiguration()
-			config.URL = fs.URL
-
-			client, err := osb.NewClient(config)
-			if err != nil {
-				t.Error(err)
-			}
-
-			actualResponse, err := client.GetCatalog()
-			if err != nil {
-				if tc.err != nil {
-					if e, a := tc.err, err; !reflect.DeepEqual(e, a) {
-						t.Errorf("Unexpected error; expected %v, got %v", e, a)
-						return
-					}
-					return
-				}
-				t.Error(err)
-				return
-			}
-
-			if e, a := tc.response, actualResponse; !reflect.DeepEqual(e, a) {
-				t.Errorf("Unexpected response\n\nExpected: %#+v\n\nGot: %#+v", e, a)
-			}
-		})
-	}
+	return conf
 }
