+++ pkg/server/update_instance_test.go
@@ -92,7 +92,7 @@ func TestUpdateInstance(t *testing.T) {
 			fs := httptest.NewServer(s.Router)
 			defer fs.Close()
 
-			config := osb.DefaultClientConfiguration()
+			config := defaultClientConfiguration()
 			config.URL = fs.URL
 
 			client, err := osb.NewClient(config)
